--- a/src/llama-model.cpp
+++ b/src/llama-model.cpp
@@ -1,4 +1,4 @@
-dee#include "llama-model.h"
+#include "llama-model.h"
 
 #include "llama-impl.h"
 #include "llama-mmap.h"