Addresses https://github.com/ggml-org/llama.cpp/pull/12801#issuecomment-2804577438

1. Permutes `q_pe` and `q_nope` before the `ggml_concat()`, then permutes `Qcur`; which will then get permuted back inside of `build_attn()` and thus should be contiguous by the time it gets to `build_attn_mha()`.
2. Performs `Vcur = ggml_cont(ctx0, Vcur)` to ensure that `Vcur` is contiguous.
3. Added back the MQA optimisation in `build_attn_mha()` (**NOTE:** requires the change in (1) to ensure `q` is contiguous!).

Apologies if it doesn't work 100% as just did it using GitHub editor as away from home, but it seems to compile and run OK for me using CUDA (not tested the performance yet though).

Leaving as a draft as @ggerganov will likely be addressing some or all of these in his subsequent MLA-related PRs - this is really just to try to track down the performance regression @fairydreaming is finding...