@jukofyork 2025-04-15
@fairydreaming If this doesn't fully resolve the performance regression then one more slightly dangerous / hacky change is to remove the `ggml_mul_mat_set_prec(kq, GGML_PREC_F32)` from here:

```
        ggml_tensor * kq = nullptr;
        if (ggml_is_contiguous(k) && ggml_is_contiguous(q) && n_head_kv == 1) {
            k = ggml_reshape_2d(ctx0, k, n_embd, n_kv);
            q = ggml_reshape_2d(ctx0, q, n_embd, n_tokens*n_head);
            kq = ggml_mul_mat(ctx0, k, q);
            kq = ggml_reshape_3d(ctx0, kq, n_kv, n_tokens, n_head);
        }
```

as IIRC, it isn't actually needed for `deepseek-r1` and your tests may have been run without this before.