@jukofyork 2025-04-15
My original tests (sorry not very scientific lol):

> I can generate over 11 tokens per second for refactoring tasks now on a machine with:
> and around 35-40 tokens per second prompt processing.

---

Master with `MLA` PR merged/ (see [here](https://github.com/ggml-org/llama.cpp/pull/12801#issuecomment-2800049058)):

```
prompt eval time =  285665.33 ms /  9848 tokens (   29.01 ms per token,    34.47 tokens per second)
       eval time = 1101768.72 ms / 11721 tokens (   94.00 ms per token,    10.64 tokens per second)
      total time = 1387434.04 ms / 21569 tokens
slot print_timing: id  0 | task 0 | 
draft acceptance rate = 0.84840 ( 9805 accepted / 11557 generated)
```

> So minimal loss of performance.

---

This PR:

```
prompt eval time =  284530.95 ms /  9846 tokens (   28.90 ms per token,    34.60 tokens per second)
       eval time = 1020527.03 ms / 11227 tokens (   90.90 ms per token,    11.00 tokens per second)
      total time = 1305057.98 ms / 21073 tokens
slot print_timing: id  0 | task 19 | 
draft acceptance rate = 0.83936 ( 9280 accepted / 11056 generated)
```

---

This PR without `ggml_mul_mat_set_prec(kq, GGML_PREC_F32)` (see [this branch](https://github.com/jukofyork/llama.cpp/tree/master-mla-optimise-q--no-f32)):

```
prompt eval time =  290181.36 ms /  9848 tokens (   29.47 ms per token,    33.94 tokens per second)
       eval time = 1076024.56 ms / 11784 tokens (   91.31 ms per token,    10.95 tokens per second)
      total time = 1366205.92 ms / 21632 tokens
slot print_timing: id  0 | task 0 | 
draft acceptance rate = 0.84582 ( 9853 accepted / 11649 generated)
```

---

So for me using CUDA, the performance drop isn't all that much (at least with this specific refactoring task using a draft model and a very high acceptance rate, 6-experts and `Q4_K` expert tensors, etc).